Fix screwed up commit ab12e7b

+++ pde/src/sgpp/pde/application/HeatEquationSolverWithStretching.cpp
@@ -33,7 +33,7 @@ HeatEquationSolverWithStretching::~HeatEquationSolverWithStretching() {
 
 void HeatEquationSolverWithStretching::constructGrid(base::Stretching& stretching, size_t level) {
   this->dim = stretching.getDimensions();
-  this->levels = static_cast<size_t>(level);
+  this->levels = static_cast<int>(level);
 
   this->myGrid = new base::LinearStretchedBoundaryGrid(stretching);
